Add files via upload
8.	Beer Kegs
Write a program that calculates the volume of n beer kegs. 
You will receive in total 3 * n lines. Every three lines will hold information for a single keg. 
First up is the model of the keg, after that is the radius of the keg, and lastly is the height of the keg.
Calculate the volume using the following formula: π * r^2 * h. 
In the end, print the model of the biggest keg.