Add files via upload
4.	Calculations
Write a program that receives a string on the first line ("add", "multiply", "subtract", "divide") and on the next two lines receives two numbers. Create four methods (for each calculation) and invoke the right one depending on the command. The method should also print the result (it needs to be void).