Add files via upload
11.	Math Operations
Write a method that receives two numbers and an operator, calculates the result, and returns it. You will be given three lines of input. The first will be the first number, the second one will be the operator, and the last one will be the second number. The possible operators are: / * + -
Print the result rounded up to the second decimal point.