Add files via upload
5.	Bomb Numbers
Write a program that reads a sequence of numbers and a special bomb number with a certain power. Your task is to detonate every occurrence of the special bomb number and according to its power - his neighbors from left and right. Detonations are performed from left to right, and all detonated numbers disappear. Finally, print the sum of the remaining elements in the sequence.